- general cleanup:   - added generics-declaration as well as @Override- and SuppressWarnings("unchecked") annotations where appropriate (i.e. implementations of Java 1.4-interface methods)   - converted space- to tab-indentation   - removed some useless null-checks, casts and unused imports   - converted implicit auto-(un)boxing statement to their respective explicit ones - CrawlerView uses URI instead of String now - fixed wrong check for serializability of argument in Field-class - crash of Paxle doesn't cause IO-Exception to be thrown because of stale temp-files in FSTempDir anymore
git-svn-id: svn://svn.paxle.net/paxle/trunk@1069 f5d3ab42-bacf-4b1a-9692-565be65a0cfe

+++ bundles/P2P/src/main/java/org/paxle/p2p/services/search/impl/SearchServerImpl.java
@@ -105,7 +105,7 @@ public class SearchServerImpl extends AServiceServer {
 			 * Process request
 			 * ================================================================ */
 		    // do a local search
-			List<ISearchResult> searchResults = this.searchProviderManager.search(queryString, queryMaxResults, queryTimeout);
+			List<ISearchResult> searchResults = this.searchProviderManager.search(queryString, queryMaxResults.intValue(), queryTimeout.longValue());
 			
 			// convert the result to XML
 			Document xmlDoc = DocumentHelper.createDocument();
